SIG-873 move v0 views, serializers, and fields.
