+++ api/app/signals/apps/signals/urls.py
@@ -1,6 +1,7 @@
 from django.urls import include, path
 
-from signals.apps.signals import routers, views
+from signals.apps.signals import routers
+from signals.apps.signals.v0 import views as v0_views
 from signals.apps.signals.v1.views import (
     GeneratePdfView,
     MainCategoryViewSet,
@@ -10,14 +11,14 @@ from signals.apps.signals.v1.views import (
 
 # API Version 0
 signal_router_v0 = routers.SignalsRouterVersion0()
-signal_router_v0.register(r'signal/image', views.SignalImageUpdateView, base_name='signal-img')
-signal_router_v0.register(r'signal', views.SignalViewSet, base_name='signal')
-signal_router_v0.register(r'auth/signal', views.SignalAuthViewSet, base_name='signal-auth')
-signal_router_v0.register(r'auth/status', views.StatusAuthViewSet, base_name='status-auth')
-signal_router_v0.register(r'auth/category', views.CategoryAuthViewSet, base_name='category-auth')
-signal_router_v0.register(r'auth/location', views.LocationAuthViewSet, base_name='location-auth')
-signal_router_v0.register(r'auth/priority', views.PriorityAuthViewSet, base_name='priority-auth')
-signal_router_v0.register(r'auth/note', views.NoteAuthViewSet, base_name='note-auth')
+signal_router_v0.register(r'signal/image', v0_views.SignalImageUpdateView, base_name='signal-img')
+signal_router_v0.register(r'signal', v0_views.SignalViewSet, base_name='signal')
+signal_router_v0.register(r'auth/signal', v0_views.SignalAuthViewSet, base_name='signal-auth')
+signal_router_v0.register(r'auth/status', v0_views.StatusAuthViewSet, base_name='status-auth')
+signal_router_v0.register(r'auth/category', v0_views.CategoryAuthViewSet, base_name='category-auth')
+signal_router_v0.register(r'auth/location', v0_views.LocationAuthViewSet, base_name='location-auth')
+signal_router_v0.register(r'auth/priority', v0_views.PriorityAuthViewSet, base_name='priority-auth')
+signal_router_v0.register(r'auth/note', v0_views.NoteAuthViewSet, base_name='note-auth')
 
 # API Version 1
 signal_router_v1 = routers.SignalsRouterVersion1()
