+++ api/app/signals/apps/signals/v1/serializers.py
@@ -5,7 +5,7 @@ from datapunt_api.rest import DisplayField, HALSerializer
 from rest_framework import serializers
 
 from signals.apps.signals.models import History, MainCategory, Signal, SubCategory
-from signals.apps.signals.serializers import _NestedDepartmentSerializer
+from signals.apps.signals.v0.serializers import _NestedDepartmentSerializer  # TODO: ../generic/.. ?
 from signals.apps.signals.v1.fields import (
     MainCategoryHyperlinkedIdentityField,
     PrivateSignalLinksField,
